cpio: fixes for Android kernel build
Allow -pd to work by changing -p from an option that takes an
argument to an option that implies there will be an argument (that
is, `-pd x` is `-p -d x` with x being the directory for -p, rather
than `-p d x` with d being the directory, as we previously interpreted
it).

Fix -d (aka --make-directories) to not be a no-op. Previously we
acted as if this was always on.

Accept --quiet and effectively just ignore it, since toybox cpio
doesn't seem to produce any output that --quiet would suppress.

Change-Id: I72d54d2beff68ff1ca088e8c2350a42f436b04f5
Signed-off-by: Jason Edson <jaysonedson@gmail.com>